services/image-service/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.3
pillow==10.3.0
pika==1.3.2
python-multipart==0.0.9

